\name{danish}
\alias{danish}
\docType{data}
\title{
Danish Fire Insurance Claims
}
\description{
These data describe large fire insurance claims in Denmark from Thursday 3rd January 1980 until Monday 31st December 1990. The data are contained in a numeric vector. They were supplied by Mette Rytgaard of Copenhagen Re}
\usage{data("danish")}
\format{
  The format is:
 atomic [1:2167] 1.68 2.09 1.73 1.78 4.61 ...
 - attr(*, "times")= POSIXt[1:2167], format: "1980-01-03 01:00:00" "1980-01-04 01:00:00" ...
}
\source{
The data is taken from package \code{evir}.
}
\examples{
data(danish)
}
