% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea-package.R
\docType{package}
\name{tea-package}
\alias{tea}
\alias{tea-package}
\title{tea: tea_a_Dataset}
\description{
This contains the data used in "Selvaratnam, R. R., de Silva, L. D., Pathmeswaran, A., & de Silva, N. R. (2003). Nutritional status and productivity of Sri Lankan tea pluckers. The Ceylon Medical Journal, 48(4), 114–118."
}
\author{
\strong{Maintainer}: Arunasalam Pathmeswaran \email{pathmes@kln.ac.lk} (\href{https://orcid.org/0000-0003-4065-2639}{ORCID})

}
\keyword{internal}
