% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\docType{data}
\name{tea}
\alias{tea}
\title{Top Tea Exporting Countries}
\format{
A data frame with 206 rows and 19 variables
\describe{
\item{Exporters}{Name of the country}
\item{2001}{2001 export value in US Dollar thousand}
\item{2002}{2001 export value in US Dollar thousand}
\item{2003}{2001 export value in US Dollar thousand}
\item{2004}{2001 export value in US Dollar thousand}
\item{2005}{2001 export value in US Dollar thousand}
\item{2006}{2001 export value in US Dollar thousand}
\item{2007}{2001 export value in US Dollar thousand}
\item{2008}{2001 export value in US Dollar thousand}
\item{2009}{2001 export value in US Dollar thousand}
\item{2010}{2001 export value in US Dollar thousand}
\item{2011}{2001 export value in US Dollar thousand}
\item{2012}{2001 export value in US Dollar thousand}
\item{2013}{2001 export value in US Dollar thousand}
\item{2014}{2001 export value in US Dollar thousand}
\item{2015}{2001 export value in US Dollar thousand}
\item{2016}{2001 export value in US Dollar thousand}
\item{2017}{2001 export value in US Dollar thousand}
\item{2018}{2001 export value in US Dollar thousand}
}
}
\source{
ITC calculations based on UN COMTRADE and ITC statistics.
}
\usage{
tea
}
\description{
Tea exporting countries (US Dollar thousand)
}
\examples{
data(tea)
head(tea)
summary(tea)
}
\keyword{datasets}
