\name{tea-package}
\alias{tea-package}
\alias{tea}
\docType{package}
\title{
Threshold Estimation Approaches
}
\description{
This package contains implementations of many of the threshold estimation approaches proposed in the literature. The estimation of the threshold is of great interest in statistics of extremes. Estimating the threshold is equivalent to choose the optimal sample fraction in tail index estimation. The sample fraction is given by \code{k/n} with \code{n} the sample size and \code{k} the number of extremes in the data or, if you wish, the exceedances over a high unknown threshold \code{u}.
}
\details{
\tabular{ll}{
Package: \tab tea\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2020-04-17\cr
License: \tab GPL-3\cr
}
}
\author{
Johannes Ossberger

Maintainer: Johannes Ossberger <johannes.ossberger@gmail.com>
}
\references{
Caeiro and Gomes (2016) <doi:10.1201/b19721-5>

Cebrian et al. (2003) <doi:10.1080/10920277.2003.10596098>

Danielsson et al. (2001) <doi:10.1006/jmva.2000.1903>

Danielsson et al. (2016) <doi:10.2139/ssrn.2717478>

De Sousa and Michailidis (2004) <doi:10.1198/106186004X12335>

Drees and Kaufmann (1998) <doi:10.1016/S0304-4149(98)00017-9>

Hall (1990) <doi:10.1016/0047-259X(90)90080-2>

Hall and Welsh (1985) <doi:10.1214/aos/1176346596>

Kratz and Resnick (1996) <doi:10.1080/15326349608807407>

Gomes et al. (2011) <doi:10.1080/03610918.2010.543297>

Gomes et al. (2012) <doi:10.1007/s10687-011-0146-6>

Gomes et al. (2013) <doi:10.1080/00949655.2011.652113>

G'Sell et al. (2016) <doi:10.1111/rssb.12122>

Guillou and Hall <doi:10.1111/1467-9868.00286>

Reiss and Thomas (2007) <doi:10.1007/978-3-0348-6336-0>

Resnick and Starica (1997) <doi:10.1017/S0001867800027889>

Thompson et al. (2009) <doi:10.1016/j.coastaleng.2009.06.003>
}